requirements.txt
autopep8==2.3.2
fastapi[standard]==0.115.12
numpy==1.26.4
torch==2.5.1
torchvision==0.20.1
Pillow==11.0.0